Start refactoring from scripts into optimize.protocol
